Improving dcd reader routines to treat large files.

+++ package/NAMDAnalyzer/lib/libFunc.pyx
@@ -10,11 +10,11 @@ np.import_array()
 
 cdef extern from "libFunc.h":
 
-    int getDCDCoor(char *fileName, long *frames, int nbrFrames, long nbrAtoms, long *selAtoms, 
-                    int selAtomsSize, long *dims, int nbrDims, int cell, long *startPos, float *outArr,
+    int getDCDCoor(char *fileName, int *frames, int nbrFrames, int nbrAtoms, int *selAtoms, 
+                    int selAtomsSize, int *dims, int nbrDims, int cell, long long *startPos, float *outArr,
                     char byteorder);
 
-    int getDCDCell(char *fileName, int *frames, int nbrFrames, int *startPos, double *outArr, char byteorder);
+    int getDCDCell(char *fileName, int *frames, int nbrFrames, long long *startPos, double *outArr, char byteorder);
 
 
     void getHBCorr(float *acceptors, int size_acceptors, int nbrFrames,
@@ -60,17 +60,17 @@ cdef extern from "libFunc.h":
 
 
 def py_getDCDCoor( fileName, 
-                np.ndarray[long, ndim=1, mode="c"] frames not None, nbrAtoms,
-                np.ndarray[long, ndim=1, mode="c"] selAtoms not None, 
-                np.ndarray[long, ndim=1, mode="c"] dims not None, cell,
-                np.ndarray[long, ndim=1, mode="c"] startPos not None, 
+                np.ndarray[int, ndim=1, mode="c"] frames not None, nbrAtoms,
+                np.ndarray[int, ndim=1, mode="c"] selAtoms not None, 
+                np.ndarray[int, ndim=1, mode="c"] dims not None, cell,
+                np.ndarray[long long, ndim=1, mode="c"] startPos not None, 
                 np.ndarray[float, ndim=3, mode="c"] outArr not None, byteorder): 
 
     res = getDCDCoor( fileName,
-                <long*> np.PyArray_DATA(frames), len(frames), nbrAtoms, 
-                <long*> np.PyArray_DATA(selAtoms), len(selAtoms),  
-                <long*> np.PyArray_DATA(dims), len(dims), cell,
-                <long*> np.PyArray_DATA(startPos),
+                <int*> np.PyArray_DATA(frames), len(frames), nbrAtoms, 
+                <int*> np.PyArray_DATA(selAtoms), len(selAtoms),  
+                <int*> np.PyArray_DATA(dims), len(dims), cell,
+                <long long*> np.PyArray_DATA(startPos),
                 <float*> np.PyArray_DATA(outArr),
                 byteorder )
 
@@ -80,13 +80,13 @@ def py_getDCDCoor( fileName,
 
 def py_getDCDCell( fileName, 
                 np.ndarray[int, ndim=1, mode="c"] frames not None, 
-                np.ndarray[int, ndim=1, mode="c"] startPos not None, 
+                np.ndarray[long long, ndim=1, mode="c"] startPos not None, 
                 np.ndarray[double, ndim=2, mode="c"] outArr not None,
                 byteorder ): 
 
     res = getDCDCell( fileName,
                 <int*> np.PyArray_DATA(frames), len(frames), 
-                <int*> np.PyArray_DATA(startPos),
+                <long long*> np.PyArray_DATA(startPos),
                 <double*> np.PyArray_DATA(outArr),
                 byteorder )
